+++ .rspec
@@ -1,0 +1,3 @@
+--color
+--format documentation
+--fail-fast
